Support mutual TLS using a certificate from a Windows cert store (#292)
Add the ability to use a client certificate located in a Windows certificate store. Previously, the client certificate and private key had to be passed by filepath or file contents. With this change, certificates and keys stored on TPM devices can be used.

Add new `windows_cert_pubsub.py` sample to show this in action.