CI Adjustment (#357)
Adjusts how CI runs so it uses OpenID. Adjusts the README for the samples so the policies shown are minimal and working in all cases, adjusts all samples so they know when they are running in CI, and adds a GitHub action that runs all samples, including Shadow, Jobs, and Fleet Provisioning. It also modifies the jobs sample to display the number of pending jobs before trying to execute/resolve them, making it like the Java V2 Jobs sample.